+++ app/src/main/java/com/triplec/triway/ListFragment.java
@@ -60,7 +60,6 @@ public class ListFragment extends Fragment {
         switch (item.getItemId()) {
             case R.id.Tabs_menu_delete:
                 //  delete all selected places
-                //boolean[] selected = adapter.getSelected();
                 SparseBooleanArray selected = adapter.getSelectedIds();
                 for(int i = selected.size() - 1; i >= 0; i--){
                     if(selected.valueAt(i)){
@@ -69,8 +68,6 @@ public class ListFragment extends Fragment {
                     }
                 }
                 adapter.removeSelection();
-                return true;
-            case R.id.Tabs_menu_edit:
                 return true;
             default:
                 return false;
@@ -92,6 +89,7 @@ public class ListFragment extends Fragment {
         TriPlan plan = builder.buildPlan();
         /*----- pseudo plan for testing -----*/
 
+        // set up list with adapter
         list = (ListView)view.findViewById(R.id.route_list);
         adapter = new PlaceListAdapter
                 (getActivity().getApplicationContext(), R.layout.fragment_list, plan.getPlaceList());
@@ -106,87 +104,5 @@ public class ListFragment extends Fragment {
         });
 
         return view;
-    }
-
-    public class PlaceListAdapter extends ArrayAdapter<TriPlace> {
-        List<TriPlace> places;
-        Context mContext;
-        LayoutInflater inflater;
-        private SparseBooleanArray mSelectedItemsIds;
-        View convert;
-
-        public PlaceListAdapter (Context context, int resourceId, List<TriPlace> places) {
-            super(context, resourceId, places);
-            this.mContext = context;
-            mSelectedItemsIds = new SparseBooleanArray();
-            inflater = LayoutInflater.from(context);
-            this.places = places;
-        }
-
-        private class ViewHolder {
-            CheckBox checkBox;
-            TextView name;
-            TextView description;
-            ImageView photo;
-        }
-
-        @Override
-        public View getView(int position, View convertView, ViewGroup parent){
-            final ViewHolder holder;
-            if (convertView == null) {
-                holder = new ViewHolder();
-                convertView = inflater.inflate(R.layout.list_item, null);
-                // Locate the TextViews in listview_item.xml
-                holder.checkBox = (CheckBox) convertView.findViewById(R.id.place_checkbox);
-                holder.name = (TextView) convertView.findViewById(R.id.place_name);
-                holder.description = (TextView) convertView.findViewById(R.id.place_description);
-                holder.photo = (ImageView) convertView.findViewById(R.id.place_photo);
-                convertView.setTag(holder);
-            } else {
-                holder = (ViewHolder) convertView.getTag();
-            }
-            // Capture position and set to the TextViews
-            holder.name.setText(places.get(position).getName());
-            holder.description.setText(places.get(position).getStreet());
-            //holder.photo.setImageResource(places.get(position).getPhoto());
-            //holder.name.setText("name " + position);
-            //holder.description.setText("description " + position);
-            holder.photo.setImageResource(R.drawable.album_city3);
-
-            holder.checkBox.setChecked(mSelectedItemsIds.get(position));
-
-            convert = convertView;
-            return convertView;
-        }
-
-        @Override
-        public void remove(TriPlace p) {
-            places.remove(p);
-            notifyDataSetChanged();
-        }
-
-        public void selectView(int position, boolean value) {
-            if (value) {
-                mSelectedItemsIds.put(position, value);
-            } else {
-                mSelectedItemsIds.delete(position);
-            }
-            notifyDataSetChanged();
-        }
-
-        public void toggleSelection(int position) {
-            selectView(position, !mSelectedItemsIds.get(position));
-            notifyDataSetChanged();
-        }
-
-        public SparseBooleanArray getSelectedIds() {
-            return mSelectedItemsIds;
-        }
-
-        public void removeSelection() {
-            mSelectedItemsIds = new SparseBooleanArray();
-            notifyDataSetChanged();
-        }
-
     }
 }
